refactor: move extension protobuf methods to separate file
