+++ src/meta/types/src/lib.rs
@@ -39,6 +39,7 @@ mod seq_value;
 mod with;
 
 mod proto_display;
+mod proto_ext;
 
 // reexport
 
@@ -140,75 +141,3 @@ pub use crate::raft_types::TypeConfig;
 pub use crate::raft_types::Vote;
 pub use crate::raft_types::VoteRequest;
 pub use crate::raft_types::VoteResponse;
-// pub use crate::raft_types::InitializeError;
-// pub use crate::raft_types::RaftChangeMembershipError;
-// pub use crate::raft_types::RaftWriteError;
-
-impl TxnCondition {
-    /// Create a txn condition that checks if the `seq` matches.
-    pub fn eq_seq(key: impl ToString, seq: u64) -> Self {
-        Self {
-            key: key.to_string(),
-            expected: ConditionResult::Eq as i32,
-            target: Some(txn_condition::Target::Seq(seq)),
-        }
-    }
-}
-
-impl TxnOp {
-    /// Create a txn operation that puts a record.
-    pub fn put(key: impl ToString, value: Vec<u8>) -> TxnOp {
-        Self::put_with_expire(key, value, None)
-    }
-
-    /// Create a txn operation that puts a record with expiration time.
-    pub fn put_with_expire(key: impl ToString, value: Vec<u8>, expire_at: Option<u64>) -> TxnOp {
-        TxnOp {
-            request: Some(txn_op::Request::Put(TxnPutRequest {
-                key: key.to_string(),
-                value,
-                prev_value: true,
-                expire_at,
-            })),
-        }
-    }
-
-    /// Create a new `TxnOp` with a `Delete` operation.
-    pub fn delete(key: impl ToString) -> Self {
-        Self::delete_exact(key, None)
-    }
-
-    /// Create a new `TxnOp` with a `Delete` operation that will be executed only when the `seq` matches.
-    pub fn delete_exact(key: impl ToString, seq: Option<u64>) -> Self {
-        TxnOp {
-            request: Some(txn_op::Request::Delete(TxnDeleteRequest {
-                key: key.to_string(),
-                prev_value: true,
-                match_seq: seq,
-            })),
-        }
-    }
-}
-
-impl TxnOpResponse {
-    /// Create a new `TxnOpResponse` of a `Delete` operation.
-    pub fn delete(key: impl ToString, success: bool, prev_value: Option<protobuf::SeqV>) -> Self {
-        TxnOpResponse {
-            response: Some(txn_op_response::Response::Delete(TxnDeleteResponse {
-                key: key.to_string(),
-                success,
-                prev_value,
-            })),
-        }
-    }
-
-    /// Create a new `TxnOpResponse` of a `Put` operation.
-    pub fn put(key: impl ToString, prev_value: Option<protobuf::SeqV>) -> Self {
-        TxnOpResponse {
-            response: Some(txn_op_response::Response::Put(TxnPutResponse {
-                key: key.to_string(),
-                prev_value,
-            })),
-        }
-    }
-}
